Jeff:  Final Commenting
Added comments explaining target and center search

Question:  Do we need to ask the robot if we have found the target then conduct the go to different points.  Currently code just runs as if it will find the center.  Performance doesn't show anything to be worried about, but it is from a coding aspect.